Responsive phototag and event map

+++ dbase/setup.sql
@@ -6,10 +6,10 @@ CREATE TABLE IF NOT EXISTS `phototags` (
   `user` INT NOT NULL,
   `username` varchar(100) NOT NULL,
   `author` varchar(100) NOT NULL,
-  `xi` int not null,
-  `yi` int not null,
-  `xii` int not null,
-  `yii` int not null,
+  `xi` float not null,
+  `yi` float not null,
+  `xii` float not null,
+  `yii` float not null,
   `added` datetime NOT NULL DEFAULT NOW(),
   `deleted` tinyint(1) DEFAULT 0
 );
